requirements.txt
openai
python-dotenv
GitPython
requests
